fix(ots): define and enforce the token-signal alignment contract
Review finding on this PR: the token setters accepted any lengths and any
mask values, and the two tests that exercised them encoded conflicting
conventions — one masked the whole prompt+completion sequence, the other only
the completion. Nothing said which was right, so a producer could serialize a
turn whose mask and log probabilities pointed at the wrong tokens and no test
would notice.

The contract, now stated on the fields themselves:

- `prompt_token_ids` is prompt-aligned and stands alone. Prompt tokens are
  never trained on, so they carry neither a mask nor log probabilities.
- `completion_token_ids` is the anchor for the completion side.
  `response_mask` and `logprobs` index into it position for position, and
  every mask entry is 0 or 1.

Enforcement at three levels, because no single one covers every path:

- `with_response_mask` rejects values outside {0, 1} — locally checkable, so
  it is checked where it is set.
- All three completion-side setters compare against whichever counterparts
  are already present, so a length disagreement trips whatever order the
  setters are called in.
- `OTSTurn::validate_token_signals` is the whole-turn check that catches what
  the setters structurally cannot: a mask or a logprob array with no
  completion tokens to align against. `TrajectoryBuilder::end_turn` runs it
  before sealing a turn, so a producer building through the builder is told
  which turn was wrong while it can still be fixed.

Both conflicting tests now use the completion-aligned convention, and the
invariants have their own cases: over-long mask, reversed setter order,
out-of-domain value, dangling mask, and a turn with no token signals at all.

The ATIF mapping doc records that `temper.response_mask` is completion-
aligned, and adds context compaction (`continued_trajectory_ref` /
`is_copied_context`) to the cannot-round-trip list — OTS models neither.

Builder tests move to `builder_test.rs` alongside the existing `*_test.rs`
convention, keeping the module under the readability limit.

Refs ARN-292
